+++ src/components/editProfile/index.js
@@ -1,2 +1,2 @@
-export { default as DeleteUser } from './DeleteUser';
+export { default as DeleteUser } from './DeleteUser/DeleteUser';
 export { EditPassword } from './EditPassword';
